relocated the protoparser away from the actual proto files so it doesn't get deleted by mistake. updated gitignore to ignore the folder holding all of the saved protos
