+++ .gitignore
@@ -58,4 +58,5 @@ ehthumbs.db
 
 # proto files #
 #####################
+src/client/components/composer/protos
 *.proto
